backend/pyproject.toml
[project]
name = "fastapi-aws"
version = "0.0.1"
description = "FastAPI AWS deploy"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "boto3>=1.39.12",
    "fastapi[standard]>=0.116.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "redis>=6.2.0",
    "sqlalchemy[asyncio]>=2.0.41",
    "web3>=7.12.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
